+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Template substitutions can now handle multiple lines, i.e. you can define custom substitutions that return a string with new line characters.
 - The "vsplit" and "hsplit" open strategies for `config.open_notes_in` will now only open a vertical/horizontal split if the window is not already split.
 
+### Fixed
+
+- Fixed URL incorrect in README.md
+
 ## [v2.3.0](https://github.com/epwalsh/obsidian.nvim/releases/tag/v2.3.0) - 2023-11-28
 
 ### Added